requirements.txt
Pillow
numpy
matplotlib
mtcnn
tensorflow==1.15